# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SIH 26128 Livestock Surveillance API",
    "description": "Checkpoint 3 livestock-health reporting, human review, privacy-preserving surveillance, and verified-only model governance API. This is a decision-support prototype, not a diagnostic system or clinically validated deployment.",
    "version": "0.3.0"
  },
  "paths": {
...
    "/api/v1/vet/cases": {
      "get": {
        "tags": [
          "reports and synchronization"
        ],
        "summary": "Vet Cases",
        "operationId": "vet_cases_api_v1_vet_cases_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Vet Cases Api V1 Vet Cases Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/vet/operational-cases": {
      "get": {
        "tags": [
          "veterinarian case management"
        ],
        "summary": "List Cases",
        "operationId": "list_cases_api_v1_vet_operational_cases_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "type": "array",
                  "title": "Response List Cases Api V1 Vet Operational Cases Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/vet/cases/{case_id}": {
      "get": {
        "tags": [
          "veterinarian case management"
        ],
        "summary": "Get Case",
        "operationId": "get_case_api_v1_vet_cases__case_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Case Api V1 Vet Cases  Case Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/vet/cases/{case_id}/assign": {
      "post": {
        "tags": [
          "veterinarian case management"
        ],
        "summary": "Assign",
        "operationId": "assign_api_v1_vet_cases__case_id__assign_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignCaseIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Assign Api V1 Vet Cases  Case Id  Assign Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/vet/cases/{case_id}/transition": {
      "post": {
        "tags": [
          "veterinarian case management"
        ],
        "summary": "Transition",
        "operationId": "transition_api_v1_vet_cases__case_id__transition_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionCaseIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Transition Api V1 Vet Cases  Case Id  Transition Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/vet/cases/{case_id}/review": {
      "post": {
        "tags": [
          "veterinarian case management"
        ],
        "summary": "Review",
        "operationId": "review_api_v1_vet_cases__case_id__review_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VetReviewIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Review Api V1 Vet Cases  Case Id  Review Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/vet/cases/{case_id}/sample-request": {
      "post": {
        "tags": [
          "veterinarian case management"
        ],
        "summary": "Sample Request",
        "operationId": "sample_request_api_v1_vet_cases__case_id__sample_request_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SampleRequestIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Sample Request Api V1 Vet Cases  Case Id  Sample Request Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/vet/cases/{case_id}/lab-referrals/{referral_id}/result": {
      "post": {
        "tags": [
          "veterinarian case management"
        ],
        "summary": "Lab Result",
        "operationId": "lab_result_api_v1_vet_cases__case_id__lab_referrals__referral_id__result_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          },
          {
            "name": "referral_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Referral Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LabResultIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Lab Result Api V1 Vet Cases  Case Id  Lab Referrals  Referral Id  Result Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/vet/cases/{case_id}/follow-ups": {
      "post": {
        "tags": [
          "veterinarian case management"
        ],
        "summary": "Follow Up",
        "operationId": "follow_up_api_v1_vet_cases__case_id__follow_ups_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FollowUpIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Follow Up Api V1 Vet Cases  Case Id  Follow Ups Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/vet/cases/{case_id}/escalate": {
      "post": {
        "tags": [
          "veterinarian case management"
        ],
        "summary": "Escalate",
        "operationId": "escalate_api_v1_vet_cases__case_id__escalate_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EscalateIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Escalate Api V1 Vet Cases  Case Id  Escalate Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssignCaseIn": {
        "properties": {
          "expected_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Version"
          },
          "veterinarian_user_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Veterinarian User Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_version"
        ],
        "title": "AssignCaseIn"
      },
...
      "EscalateIn": {
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 300,
            "minLength": 3,
            "title": "Reason"
          },
          "level": {
            "type": "integer",
            "maximum": 3.0,
            "minimum": 1.0,
            "title": "Level"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "reason",
          "level"
        ],
        "title": "EscalateIn"
      },
...
      "FollowUpIn": {
        "properties": {
          "follow_up_type": {
            "type": "string",
            "enum": [
              "PHONE",
              "FIELD_VISIT",
              "TREATMENT_CHECK",
              "OTHER"
            ],
            "title": "Follow Up Type"
          },
          "notes": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 2,
            "title": "Notes"
          },
          "due_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "follow_up_type",
          "notes"
        ],
        "title": "FollowUpIn"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LabResultIn": {
        "properties": {
          "expected_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Version"
          },
          "result_status": {
            "type": "string",
            "enum": [
              "CONFIRMED",
              "NEGATIVE",
              "INCONCLUSIVE"
            ],
            "title": "Result Status"
          },
          "result_label": {
            "type": "string",
            "maxLength": 160,
            "minLength": 2,
            "title": "Result Label"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_version",
          "result_status",
          "result_label"
        ],
        "title": "LabResultIn"
      },
...
      "SampleRequestIn": {
        "properties": {
          "expected_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Version"
          },
          "sample_identifier": {
            "type": "string",
            "maxLength": 120,
            "minLength": 4,
            "pattern": "^[A-Za-z0-9._-]+$",
            "title": "Sample Identifier"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_version",
          "sample_identifier"
        ],
        "title": "SampleRequestIn"
      },
...
      "TransitionCaseIn": {
        "properties": {
          "expected_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Version"
          },
          "to_state": {
            "type": "string",
            "enum": [
              "UNDER_REVIEW",
              "LAB_PENDING",
              "CLOSED"
            ],
            "title": "To State"
          },
          "reason": {
            "type": "string",
            "maxLength": 300,
            "minLength": 3,
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_version",
          "to_state",
          "reason"
        ],
        "title": "TransitionCaseIn"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VetReviewIn": {
        "properties": {
          "expected_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Version"
          },
          "outcome": {
            "type": "string",
            "enum": [
              "CONFIRMED",
              "CORRECTED",
              "INCONCLUSIVE"
            ],
            "title": "Outcome"
          },
          "verified_label": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 160,
                "minLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified Label"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_version",
          "outcome"
        ],
        "title": "VetReviewIn"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}